Organize files with updated method
